3.fastapi_advanced/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
starlette==0.14.2
SQLAlchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
requests==2.26.0
tenacity==8.0.1
python-dotenv==0.19.0
